Add md5 checksum for dataset and move demo scripts under f3rm
