+++ f3rm/scripts/demo_clip_features.py
@@ -9,12 +9,9 @@ from f3rm.features.clip import tokenize
 from f3rm.features.clip_extract import CLIPArgs, extract_clip_features
 
 _MODULE_DIR = os.path.dirname(os.path.abspath(__file__))
-_IMAGE_DIR = os.path.join(_MODULE_DIR, "images")
+_IMAGE_DIR = os.path.join(_MODULE_DIR, "../../scripts/images")
 
-image_paths = [
-    os.path.join(_IMAGE_DIR, name)
-    for name in ["frame_1.png", "frame_2.png", "frame_3.png"]
-]
+image_paths = [os.path.join(_IMAGE_DIR, name) for name in ["frame_1.png", "frame_2.png", "frame_3.png"]]
 
 
 @torch.no_grad()
